+++ src/app/reply/reply.component.ts
@@ -25,6 +25,7 @@ export class ReplyComponent implements OnInit {
   position = new FormControl(this.positionOption);
   showDelay = new FormControl(50);
   hideDelay = new FormControl(50);
+  busyRequest: Boolean = false;
 
   /* The Authorized User (The one who made Log in) */
   authorizedUser: string = localStorage.getItem("username");
@@ -56,7 +57,7 @@ export class ReplyComponent implements OnInit {
 
   /**
    * close popup when another nested popup appear and open the new popup
-   * @param kweek 
+   * @param kweek
    * No @returns
    */
   nestedDialog(kweek: Kweek): void {
@@ -81,7 +82,11 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   like(kweek: Kweek): void {
-    this.kweekFunc.like(kweek);
+    if(!this.busyRequest) {
+      this.busyRequest = true;
+      this.kweekFunc.like(kweek);
+      this.busyRequest = false;
+    }
   }
 
   /**
@@ -90,7 +95,11 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   unlike(kweek: Kweek): void {
-    this.kweekFunc.unlike(kweek);
+    if(!this.busyRequest) {
+      this.busyRequest = true;
+      this.kweekFunc.unlike(kweek);
+      this.busyRequest = false;
+    }
   }
 
   /**
@@ -99,7 +108,11 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   rekweek(kweek: Kweek): void {
-    this.kweekFunc.rekweek(kweek);
+    if(!this.busyRequest) {
+      this.busyRequest = true;
+      this.kweekFunc.rekweek(kweek);
+      this.busyRequest = false;
+    }
   }
 
   /**
@@ -108,7 +121,11 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   unrekweek(kweek: Kweek): void {
-    this.kweekFunc.unrekweek(kweek);
+    if (!this.busyRequest) {
+      this.busyRequest = true;
+      this.kweekFunc.unrekweek(kweek);
+      this.busyRequest = false;
+    }
   }
 
   /**
@@ -117,9 +134,13 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   deleteRoot_ClickedKweek(kweek: Kweek): void {
-    this.kweekService.deleteKweek(kweek.id).subscribe(() => {
-      this.dialogRef.close();
-    });
+    if (!this.busyRequest) {
+      this.busyRequest = true;
+      this.kweekService.deleteKweek(kweek.id).subscribe(() => {
+        this.dialogRef.close();
+        this.busyRequest = false;
+      });
+    }
   }
 
   /**
@@ -128,9 +149,13 @@ export class ReplyComponent implements OnInit {
    * No @returns
    */
   deleteReply(kweek: Kweek): void {
-    this.kweekService.deleteKweek(kweek.id).subscribe(() => {
-      const indexToDelete = this.replies.indexOf(kweek);
-      this.replies.splice(indexToDelete, 1);
-    });
+    if (!this.busyRequest) {
+      this.busyRequest = true;
+      this.kweekService.deleteKweek(kweek.id).subscribe(() => {
+        const indexToDelete = this.replies.indexOf(kweek);
+        this.replies.splice(indexToDelete, 1);
+        this.busyRequest = false;
+      });
+    }
   }
 }
